requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
retrying==1.3.4
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
